+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.8] - 2025-10-28
+
+### Added
+
+- add UI for recording
+
+### Changed
+
+- Remove global variable Output path for cli param
+- Mode cli param to model
+
 ---
 
 ## [1.0.7] - 2025-10-20